requirements.txt
markdown-it-py
watchdog
trio
selenium
